+++ ui/tabs/plots_tab/__init__.py
@@ -1,0 +1,1 @@
+"""Plots tab modules split by category."""