+++ ui/tabs/plots_tab/plots_happiness_tab.py
@@ -1,0 +1,13 @@
+"""Happiness plot tab rendering."""
+
+import os
+
+from ui.tabs.plots_tab.plots_base_tab import PlotsBaseTabMixin
+
+
+class HappinessPlotsTabMixin(PlotsBaseTabMixin):
+    def _add_happiness_plot_tab(self, png_file):
+        """Render the happiness plot tab."""
+        filename = os.path.basename(png_file)
+        tab_name = os.path.splitext(filename)[0].replace("_", " ").title()
+        self._add_plot_image_tab(png_file, tab_name)